+++ src/arguments.rs
@@ -1,5 +1,10 @@
-use crate::io::{FileType, WriteType};
-use clap::{crate_authors, value_parser, Arg, ArgAction, ArgMatches, Command};
+use rustc_hash::FxHashMap;
+
+use crate::{
+    errors::ArgumentError,
+    io::{FileType, WriteType},
+};
+use std::fmt::{Debug, Display};
 
 /// Indicates how many reference files are passed
 #[derive(Clone)]
@@ -12,122 +17,586 @@ pub enum Reference {
     None,
 }
 
-/// Create a container for dealing with clap and being able to test arg parsing.
-pub enum ClapApp {}
+enum DefaultValue {
+    None,
+    Int(usize),
+    Float(f64),
+}
 
-impl ClapApp {
-    /// Create and return the clap::App
-    pub fn get() -> Command {
-        Command::new("Multi-threaded Bader Charge Analysis")
-            .author(crate_authors!())
-            .version("0.4.5")
-            .arg(Arg::new("file")
-                .required(true)
-                .index(1)
-                .help("The file to analyse."))
-            .arg(Arg::new("output")
-                .short('o')
-                .long("output")
-                .action(ArgAction::SetTrue)
-                .help("Output the Bader atoms.")
-                .long_help(
-"Output the Bader atoms in the same file format as the input density.
-This can be used in conjunction with the index flag to specify a
-specific atom. Without the index flag it will print all the atoms."))
-            .arg(Arg::new("index")
-                .short('i')
-                .long("index")
-                .action(ArgAction::Append)
-                .requires("output")
-                .value_parser(1..)
-                .help("Index of Bader atoms to be written out.")
-                .long_help(
-"An index of a Bader atom to be written out, starting at 1. This flag
-requires the output flag to be set. Multiple atoms can be written by
-repeating the flag ie. bca CHGCAR -oi 1 -i 2."))
-            .arg(Arg::new("file type")
-                .short('t')
-                .long("type")
-                .num_args(1)
-                .value_parser(["cube", "vasp"])
-                .ignore_case(false)
-                .help("The file type of the charge density.")
-                .long_help(
-"The file type of the input file. If this is not supplied the type will attempt
-to be infered from the filename."))
-            .arg(Arg::new("reference")
-                .short('r')
-                .long("ref")
-                .action(ArgAction::Append)
-                .num_args(1)
-                .value_parser(value_parser!(String))
-                .help("File(s) containing reference charge.")
-                .long_help(
-"A reference charge to do the partitioning upon. Two files can be passed
-by using multiple flags (bca CHGCAR -r AECCAR0 -r AECCAR2). If two files are
-passed they are summed together."))
-            .arg(Arg::new("spin")
-                .short('s')
-                .long("spin")
-                .num_args(1)
-                .value_parser(value_parser!(String))
-                .help("File containing spin density.")
-                .long_help(
-"A path to the spin density associated with the original file. This is primarily
-for cube files as if spin density exists in a CHGCAR it will be read automatically.
-If using with VASP outputs then the files for charge and spin density must only
-contain a single density (ie. the original file has been split)."))
-            .arg(Arg::new("all electron")
-                .short('a')
-                .long("aec")
-                .help("Convience flag for reading both aeccars.")
-                .action(ArgAction::SetTrue)
-                .conflicts_with("reference"))
-            .arg(Arg::new("vacuum tolerance")
-                .short('v')
-                .long("vac")
-                .default_value("1E-6")
-                .value_parser(value_parser!(f64))
-                .num_args(1)
-                .help("Cut-off at which charge is considered vacuum.")
-                .long_help(
-"Values of density below the supplied value are considered vacuum and are not
-included in the calculation."))
-            .arg(Arg::new("weight tolerance")
-                .short('w')
-                .long("weight")
-                .default_value("1E-8")
-                .value_parser(value_parser!(f64))
-                .num_args(1)
-                .help("Cut-off at which contributions to the weighting will be ignored.")
-                .long_help(
-"Values of density below the supplied value are ignored from the weighting and
-included in the calculation. By raising the tolerance the calculation speed can
-be increased but every ignored weight is unaccounted for in the final partitions.
-Be sure to test this!"))
-            .arg(Arg::new("maximum distance")
-                .short('m')
-                .long("max_dist")
-                .default_value("0.1")
-                .value_parser(value_parser!(f64))
-                .num_args(1)
-                .help("Cut-off after which an error will be thrown for distance of Bader maximum to atom.")
-                .long_help(
-"The distance allowed between Bader maximum and its associated atom, in angstrom, before
-an error is thrown. This will cause a hard crash of the program, consider whether
-increasing the cut-off or adding a \"ghost atom\" at the location of the Bader
-maximum is more appropriate."))
-            .arg(Arg::new("threads")
-                .short('J')
-                .long("threads")
-                .num_args(1)
-                .default_value("0")
-                .value_parser(value_parser!(usize))
-                .help("Number of threads to distribute the calculation over.")
-                .long_help(
-"The number of threads to be used by the program. A default value of 0 is used
-to allow the program to best decide how to use the available hardware. It does
-this by using the minimum value out of the number cores available and 12."))
+enum AllowedValue {
+    None,
+    Strs(Vec<String>),
+}
+
+struct Opt {
+    name: String,
+    short_help: String,
+    long_help: String,
+    short_flag: String,
+    long_flag: String,
+    takes_value: bool,
+    multiple_values: bool,
+    default_value: DefaultValue,
+    allowed_values: AllowedValue,
+}
+
+pub struct App {
+    options: Vec<Opt>,
+    help: Opt,
+    max_width: usize,
+}
+
+impl App {
+    pub fn new() -> Self {
+        //        let arg =
+        //            Opt::new("file",
+        //                     "The file to analyse.",
+        //                     "The file containing the charge density to analyse.",
+        //                     None,
+        //                     None,
+        //                     true,
+        //                     DefaultValue::None,
+        //                     AllowedValue::None);
+        // Start options
+        let options = vec![
+            Opt{
+                name: String::from("aec"),
+                short_help: String::from("Convience flag for reading both AECCARs."),
+                long_help: String::from("
+\tFlag for reading and summing both the AECCAR0 and AECCAR2 from a VASP calculation"),
+                short_flag: String::from("a"),
+                long_flag: String::from("aec"),
+                takes_value: false,
+                multiple_values: false,
+                default_value: DefaultValue::None,
+                allowed_values: AllowedValue::None,
+            },
+            Opt {
+                name: String::from("file type"),
+                short_help: String::from("File containing spin density."),
+                long_help: String::from("
+\tA path to the spin density associated with the original file. This is primarily
+\tfor cube files as if spin density exists in a CHGCAR it will be read automatically.
+\tIf using with VASP outputs then the files for charge and spin density must only
+\tcontain a single density (ie. the original file has been split)."),
+                short_flag: String::from("f"),
+                long_flag: String::from("file_type"),
+                takes_value: true,
+                multiple_values: false,
+                default_value: DefaultValue::None,
+                allowed_values: AllowedValue::Strs(vec![String::from("cube"), String::from("vasp")]),
+            },
+            Opt {
+                name: String::from("index"),
+                short_help: String::from("Index of Bader atoms to be written out."),
+                long_help: String::from("
+\tAn index of a Bader atom to be written out, starting at 1. This flag requires
+\tthe output flag to be set. Multiple atoms can only be written by repeating the
+\tflag ie. bca CHGCAR -oi 1 -i 2."),
+                short_flag: String::from("i"),
+                long_flag: String::from("index"),
+                takes_value: true,
+                multiple_values: true,
+                default_value: DefaultValue::None,
+                allowed_values: AllowedValue::None,
+            },
+            Opt {
+                name: String::from("maximum distance"),
+                short_help: String::from("Cut-off after which an error will be thrown for distance of Bader maximum to atom."),
+                long_help: String::from("
+\tThe distance allowed between Bader maximum and its associated atom, in angstrom, before
+\tan error is thrown. This will cause a hard crash of the program, consider whether
+\tincreasing the cut-off or adding a \"ghost atom\" at the location of the Bader
+\tmaximum is more appropriate."),
+                short_flag: String::from("m"),
+                long_flag: String::from("max_dist"),
+                takes_value: true,
+                multiple_values: false,
+                default_value: DefaultValue::Float(0.1),
+                allowed_values: AllowedValue::None,
+            },
+            Opt {
+                name: String::from("output"),
+                short_help: String::from("Output the Bader atoms."),
+                long_help: String::from("
+\tOutput the bader atoms in the same file format as the input density.
+\tthis can be used in conjunction with the index flag to specify a
+\tspecific atom. without the index flag it will print all the atoms."),
+                short_flag: String::from("o"),
+                long_flag: String::from("output"),
+                takes_value: false,
+                multiple_values: false,
+                default_value: DefaultValue::None,
+                allowed_values: AllowedValue::None,
+            },
+            Opt {
+                name: String::from("reference"),
+                short_help: String::from("File(s) containing reference charge."),
+                long_help: String::from("
+\tA reference charge to do the partitioning upon. Two files can be passed
+\tby using multiple flags (bca CHGCAR -r AECCAR0 -r AECCAR2). If two files are
+\tpassed they are summed together."),
+                short_flag: String::from("r"),
+                long_flag: String::from("ref"),
+                takes_value: true,
+                multiple_values: true,
+                default_value: DefaultValue::None,
+                allowed_values: AllowedValue::None,
+            },
+            Opt {
+                name: String::from("spin"),
+                short_help: String::from("File containing spin density."),
+                long_help: String::from("
+\tA path to the spin density associated with the original file. This is primarily
+\tfor cube files as if spin density exists in a CHGCAR it will be read automatically.
+\tIf using with VASP outputs then the files for charge and spin density must only
+\tcontain a single density (ie. the original file has been split)."),
+                short_flag: String::from("s"),
+                long_flag: String::from("spin"),
+                takes_value: true,
+                multiple_values: false,
+                default_value: DefaultValue::None,
+                allowed_values: AllowedValue::None,
+            },
+            Opt {
+                name: String::from("threads"),
+                short_help: String::from("Number of threads to distribute the calculation over."),
+                long_help: String::from("
+\tThe number of threads to be used by the program. A default value of 0 is used
+\tto allow the program to best decide how to use the available hardware. It does
+\tthis by using the minimum value out of the number cores available and 12."),
+                short_flag: String::from("t"),
+                long_flag: String::from("threads"),
+                takes_value: true,
+                multiple_values: false,
+                default_value: DefaultValue::Int(0),
+                allowed_values: AllowedValue::None,
+            },
+            Opt {
+                name: String::from("vacuum tolerance"),
+                short_help: String::from("Cut-off at which charge is considered vacuum."),
+                long_help: String::from("
+\tValues of density below the supplied value are considered vacuum and are not
+\tincluded in the calculation."),
+                short_flag: String::from("v"),
+                long_flag: String::from("vac"),
+                takes_value: true,
+                multiple_values: false,
+                default_value: DefaultValue::Float(1E-6),
+                allowed_values: AllowedValue::None,
+            },
+            Opt {
+                name: String::from("weight tolerance"),
+                short_help: String::from("Cut-off at which contributions to the weighting will be ignored."),
+                long_help: String::from("
+\tValues of density below the supplied value are ignored from the weighting and
+\tincluded in the calculation. By raising the tolerance the calculation speed can
+\tbe increased but every ignored weight is unaccounted for in the final partitions.
+\tBe sure to test this!"),
+                short_flag: String::from("w"),
+                long_flag: String::from("weight"),
+                takes_value: true,
+                multiple_values: false,
+                default_value: DefaultValue::Float(1E-8),
+                allowed_values: AllowedValue::None,
+            }
+        ];
+        // End options
+        let help =
+            Opt { name: String::from("help"),
+                  short_help:
+                      String::from("Print help (see more with '--help')"),
+                  long_help:
+                      String::from("Print help (see a summary with '-h')"),
+                  short_flag: String::from("h"),
+                  long_flag: String::from("help"),
+                  takes_value: false,
+                  multiple_values: false,
+                  default_value: DefaultValue::None,
+                  allowed_values: AllowedValue::None };
+        let max_width =
+            options.iter()
+                   .map(|o| {
+                       o.short_flag.len()
+                       + o.long_flag.len()
+                       + if o.takes_value { 7 + o.name.len() } else { 4 }
+                   })
+                   .max()
+                   .unwrap_or(0);
+        Self { options,
+               help,
+               max_width }
+    }
+
+    fn get_options_text(&self, long: bool) -> String {
+        self.options
+            .iter()
+            .map(|o| {
+                format!(" {:<width$}{}{}{}\n",
+                        // the short and long flag
+                        if o.takes_value {
+                            format!("-{},--{} [{}]",
+                                    o.short_flag, o.long_flag, o.name)
+                        } else {
+                            format!("-{},--{}", o.short_flag, o.long_flag)
+                        },
+                        // the help text
+                        if long {
+                            o.long_help.to_string()
+                        } else {
+                            format!("  {}", o.short_help)
+                        },
+                        // does it have a default value
+                        match o.default_value {
+                            DefaultValue::None =>
+                                if long {
+                                    String::from("\n")
+                                } else {
+                                    String::with_capacity(0)
+                                },
+                            DefaultValue::Int(u) =>
+                                if long {
+                                    format!("\n\n\t[default: {}]\n", u)
+                                } else {
+                                    format!(" [default: {}]", u)
+                                },
+                            DefaultValue::Float(f) =>
+                                if long {
+                                    format!("\n\n\t[default: {}]\n", f)
+                                } else {
+                                    format!(" [default: {}]", f)
+                                },
+                        },
+                        // does it have allowed values
+                        match &o.allowed_values {
+                            AllowedValue::None => String::with_capacity(0),
+                            AllowedValue::Strs(v) =>
+                                if long {
+                                    format!("\n\t[possible values: {:?}]\n", v)
+                                } else {
+                                    format!(" [possible values: {:?}]", v)
+                                },
+                        },
+                        width = self.max_width)
+            })
+            .collect()
+    }
+
+    fn get_help_text(&self, long: bool) -> String {
+        format!(" {:<width$}{}",
+                format!("-{},--{}", self.help.short_flag, self.help.long_flag),
+                if long {
+                    format!("\n\t{}", self.help.long_help)
+                } else {
+                    format!("  {}", self.help.short_help)
+                },
+                width = self.max_width)
+    }
+
+    fn get_option_from_short_flag(&self,
+                                  f: String)
+                                  -> Result<&Opt, ArgumentError> {
+        if f == self.help.short_flag {
+            return Err(ArgumentError::ShortHelp(self));
+        }
+        for opt in self.options.iter() {
+            if f == opt.short_flag {
+                return Ok(opt);
+            }
+        }
+        Err(ArgumentError::NotFlag(f))
+    }
+    fn get_option_from_long_flag(&self,
+                                 f: String)
+                                 -> Result<&Opt, ArgumentError> {
+        if f == self.help.long_flag {
+            return Err(ArgumentError::LongHelp(self));
+        }
+        for opt in self.options.iter() {
+            if f == opt.long_flag {
+                return Ok(opt);
+            }
+        }
+        Err(ArgumentError::NotFlag(f))
+    }
+
+    pub fn parse_args(&self, args: Vec<&str>) -> Result<Args, ArgumentError> {
+        let mut arguments = FxHashMap::<String, String>::default();
+        let mut multi_arguments = FxHashMap::<String, Vec<String>>::default();
+        args.iter().enumerate()
+                   .try_for_each(|(i, flag)| {
+                       match flag.strip_prefix('-') {
+                           // starts with - so is arg
+                           Some(stripped_flag) => {
+                               // longer than 1 char after being stripped means either cluster
+                               // of flags or long flag
+                               match stripped_flag.len().cmp(&1) {
+                                   std::cmp::Ordering::Greater => match stripped_flag.strip_prefix('-') {
+                                       // matches - again is long flag
+                                       Some(long_flag) => {
+                                           match self.get_option_from_long_flag(long_flag.to_string()) {
+                                               // if it's in the argument list we are good and
+                                               // just need to check if it takes values
+                                               Ok(o) => {if o.takes_value {
+                                                   let value = match args.get(i + 1) {
+                                                       Some(v) => v.to_string(),
+                                                       None => return Err(ArgumentError::NoValue(long_flag.to_string())),
+                                                   };
+                                                   if o.multiple_values {
+                                                       let key = multi_arguments.entry(o.name.to_string()).or_insert(vec![]);
+                                                       key.push(value);
+                                                   } else {
+                                                       arguments.insert(o.name.to_string(), value);
+                                                   }
+                                               } else {
+                                                   arguments.insert(o.name.to_string(), String::with_capacity(0));
+                                               };
+                                               Ok(())},
+                                               // else it isn't an argument
+                                               Err(e) => Err(e),
+                                           }
+                                       },
+                                       // doesn't match again so is cluster of short flags
+                                       None => {
+                                           stripped_flag.chars().try_for_each(|f| match self.get_option_from_short_flag(f.to_string()) {
+                                               // if it's in the argument list we are good and
+                                               // just need to check if it takes values
+                                               Ok(o) => {if o.takes_value {
+                                                   let value = match args.get(i + 1) {
+                                                       Some(v) => v.to_string(),
+                                                       None => return Err(ArgumentError::NoValue(f.to_string())),
+                                                   };
+                                                   if o.multiple_values {
+                                                       let key = multi_arguments.entry(o.name.to_string()).or_insert(vec![]);
+                                                       key.push(value);
+                                                   } else {
+                                                       arguments.insert(o.name.to_string(), value);
+                                                   }
+                                               } else {
+                                                   arguments.insert(o.name.to_string(), String::with_capacity(0));
+                                               };
+                                               Ok(())},
+                                               // else it isn't an argument
+                                               Err(e) => Err(e),
+                                           })
+                                       },
+                                   },
+                               // is 1 in length so short flag
+                               std::cmp::Ordering::Equal => match self.get_option_from_short_flag(stripped_flag.to_string()) {
+                                       // if it's in the argument list we are good and just
+                                       // need to check if it takes values
+                                       Ok(o) => {if o.takes_value {
+                                           let value = match args.get(i + 1) {
+                                               Some(v) => v.to_string(),
+                                               None => return Err(ArgumentError::NoValue(stripped_flag.to_string())),
+                                           };
+                                           if o.multiple_values {
+                                               let key = multi_arguments.entry(o.name.to_string()).or_insert(vec![]);
+                                               key.push(value);
+                                           } else {
+                                               arguments.insert(o.name.to_string(), value);
+                                           }
+                                       } else {
+                                           arguments.insert(o.name.to_string(), String::with_capacity(0));
+                                       };
+                                       Ok(())},
+                                       // else it isn't an argument
+                                       Err(e) => Err(e),
+                                   },
+                               std::cmp::Ordering::Less => Err(ArgumentError::NotFlag("-".to_string()))
+                               }
+                           },
+                           None => Ok(()),
+                       }
+                   })?;
+        let mut file = String::with_capacity(0);
+        args.windows(2).for_each(|window| {
+            let flag = window[0].to_string();
+            let arg = window[1].to_string();
+            // doesn't start with - so could be out file
+            if !arg.starts_with('-') {
+                match flag.strip_prefix('-') {
+                    // previous argument does start with - so we need to check if it is a passed value
+                    Some(short_flag) => {
+                        match short_flag.strip_prefix('-') {
+                            Some(long_flag) => if !self.get_option_from_long_flag(long_flag.to_string()).unwrap().takes_value {
+                                file = arg;
+                            },
+                            None => if short_flag.len() > 1 {
+                                let mut takes_value_flag = false;
+                                for f in short_flag.chars() {
+                                    if self.get_option_from_short_flag(f.to_string()).unwrap().takes_value {
+                                        takes_value_flag = true;
+                                        break
+                                    }
+                                }
+                                if !takes_value_flag {
+                                    file = arg;
+                                }
+                            } else if !self.get_option_from_short_flag(short_flag.to_string()).unwrap().takes_value {
+                                file = arg;
+                            },
+                        }
+                    },
+                    None => file = arg,
+                }
+            }
+        });
+        let file_type = match arguments.get("file type") {
+            Some(ftype) => {
+                let ftype = ftype.to_lowercase();
+                if ftype.eq("cube") {
+                    FileType::Cube
+                } else if ftype.eq("vasp") {
+                    FileType::Vasp
+                } else {
+                    return Err(ArgumentError::NotValidValue(String::from("file type"), ftype));
+                }
+            }
+            None => {
+                let f = file.to_lowercase();
+                if f.contains("cube") {
+                    FileType::Cube
+                } else if f.contains("car") {
+                    FileType::Vasp
+                } else {
+                    eprintln!("Cannot detect file type, attempting to read as VASP.");
+                    FileType::Vasp
+                }
+            }
+        };
+        let weight_tolerance = match arguments.get("weight tolerance") {
+            Some(w) => match w.parse::<f64>() {
+                Ok(f) => f,
+                Err(_) => return Err(ArgumentError::Unparsable(String::from("weight tolerance"), w.to_string(), String::from("f64")))
+            },
+            None => match self.get_option_from_short_flag(String::from("w")).unwrap().default_value {
+                DefaultValue::Float(f) => f,
+                _ => panic!(""),
+            }
+        };
+        let maximum_distance = match arguments.get("maximum distance") {
+            Some(m) => match m.parse::<f64>() {
+                Ok(f) => f,
+                Err(_) => return Err(ArgumentError::Unparsable(String::from("maximum distance"), m.to_string(), String::from("f64")))
+            },
+            None => match self.get_option_from_short_flag(String::from("m")).unwrap().default_value {
+                DefaultValue::Float(f) => f,
+                _ => panic!(""),
+            }
+        };
+        let vacuum_tolerance = match arguments.get("vacuum tolerance") {
+            Some(m) => {
+                if m.to_lowercase() == "none" {
+                    None
+                } else {
+                    match m.parse::<f64>() {
+                    Ok(f) => Some(f),
+                    Err(_) => return Err(ArgumentError::Unparsable(String::from("vacuum tolerance"), m.to_string(), String::from("f64")))
+                }
+                }
+            }
+            None => match self.get_option_from_short_flag(String::from("v"))
+                              .unwrap()
+                              .default_value
+            {
+                DefaultValue::Float(f) => Some(f),
+                _ => panic!(""),
+            },
+        };
+        let mut threads = match arguments.get("threads") {
+            Some(t) => match t.parse::<usize>() {
+                Ok(f) => f,
+                Err(_) => return Err(ArgumentError::Unparsable(String::from("threads"), t.to_string(), String::from("usize")))
+            },
+            None => match self.get_option_from_short_flag(String::from("t")).unwrap().default_value {
+                DefaultValue::Int(u) => u,
+                _ => panic!(""),
+            }
+        };
+        if threads == 0 {
+            threads = num_cpus::get().min(12);
+        }
+        let output = match arguments.get("output") {
+            Some(_) => {
+                match multi_arguments.get("index") {
+                    Some(v) => {
+                        let mut a = Vec::with_capacity(v.len());
+                        for i in v.iter() {
+                            match i.parse::<isize>() {
+                                Ok(i) => a.push(i - 1),
+                                Err(_) => return Err(ArgumentError::Unparsable(String::from("index"), i.to_string(), String::from("isize")))
+                            }
+                        }
+                        WriteType::Atom(a)
+                    },
+                    None => WriteType::Atom(Vec::with_capacity(0)),
+                }
+            },
+            None => match multi_arguments.get("index") {
+                Some(_) => return Err(ArgumentError::MissingDependant(String::from("index"), String::from("output"))),
+                None => WriteType::None,
+            }
+        };
+        let reference = match multi_arguments.get("reference") {
+            Some(v) => match v.len().cmp(&2) {
+                std::cmp::Ordering::Less => Reference::One(v[0].to_string()),
+                std::cmp::Ordering::Equal => Reference::Two(v[0].to_string(), v[1].to_string()),
+                std::cmp::Ordering::Greater => return Err(ArgumentError::TooManyValues(String::from("reference"), 2, v.len())),
+            }
+            None => match arguments.get("aec") {
+                Some(_) => match file_type {
+                    FileType::Vasp => Reference::Two(String::from("AECCAR0"), String::from("AECCAR2")),
+                    FileType::Cube => return Err(ArgumentError::WrongFileType(String::from("aec"), String::from("cube"))),
+                },
+                None => Reference::None,
+            }
+        };
+        let spin = arguments.get("spin").cloned();
+        if file.is_empty() {
+            return Err(ArgumentError::NoFile(self));
+        }
+        Ok(Args { file,
+                  file_type,
+                  weight_tolerance,
+                  maximum_distance,
+                  output,
+                  reference,
+                  spin,
+                  threads,
+                  vacuum_tolerance })
+    }
+}
+
+impl Default for App {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Display for App {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let usage = String::from("Usage: bca [OPTIONS] <file>");
+        let arguments =
+            String::from("Arguments:\n <file>  The file to analyse.");
+        let options = self.get_options_text(false);
+        let help = self.get_help_text(false);
+        write!(f,
+               "{}\n\n{}\n\nOptions:\n{}{}",
+               usage, arguments, options, help)
+    }
+}
+
+impl Debug for App {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let usage = String::from("Usage: bca [OPTIONS] <file>");
+        let arguments =
+            String::from("Arguments:\n <file>  The file containing the charge density to analyse.");
+        let options = self.get_options_text(true);
+        let help = self.get_help_text(true);
+        write!(f,
+               "{}\n\n{}\n\nOptions:\n{}{}",
+               usage, arguments, options, help)
     }
 }
 
@@ -153,162 +622,67 @@ pub struct Args {
     pub vacuum_tolerance: Option<f64>,
 }
 
-impl Args {
-    /// Initialises the structure from the command-line arguments.
-    pub fn new(arguments: ArgMatches) -> Self {
-        // Collect file
-        let file = arguments.get_one::<String>("file").unwrap().to_string();
-        // Collect write charge info
-        let output = if arguments.get_flag("output") {
-            let atoms: Vec<isize> = match arguments.get_many::<i64>("index") {
-                Some(vec) => vec.into_iter().map(|i| *i as isize - 1).collect(),
-                None => Vec::with_capacity(0),
-            };
-            WriteType::Atom(atoms)
-        } else {
-            WriteType::None
-        };
-
-        // Collect file type
-        let file_type = arguments.get_one::<String>("file type");
-        let file_type = match file_type {
-            Some(ftype) => {
-                if ftype.eq("cube") {
-                    FileType::Cube
-                } else {
-                    FileType::Vasp
-                }
-            }
-            None => {
-                if file.to_lowercase().contains("cube") {
-                    FileType::Cube
-                } else if file.to_lowercase().contains("car") {
-                    FileType::Vasp
-                } else {
-                    println!("Error: File-type cannot be infered, attempting to read as VASP");
-                    FileType::Vasp
-                }
-            }
-        };
-        // Collect weight tolerance
-        // safe to unwrap as threads has a default value of 1E-8
-        let weight_tolerance = arguments.get_one::<f64>("weight tolerance")
-                                        .unwrap()
-                                        .max(1E-13);
-        // Collect maxima tolerance
-        // safe to unwrap as threads has a default value of 0.1
-        let maximum_distance =
-            *arguments.get_one::<f64>("maximum distance").unwrap();
-        // Collect threads
-        // safe to unwrap as threads has a default value of 0
-        let threads = match *arguments.get_one::<usize>("threads").unwrap() {
-            0 => num_cpus::get().min(12),
-            x => x,
-        };
-        // Collect vacuum tolerance
-        let vacuum_tolerance =
-            arguments.get_one::<f64>("vacuum tolerance").copied();
-        // Collect reference files
-        let references: Vec<_> = if arguments.get_flag("all electron") {
-            if let FileType::Vasp = file_type {
-                vec!["AECCAR0", "AECCAR2"]
-            } else {
-                panic!("Error: Cannot use AECCAR flag for non VASP file-types.")
-            }
-        } else {
-            match arguments.get_many::<String>("reference") {
-                Some(x) => x.map(|s| s.as_str()).collect(),
-                None => Vec::with_capacity(0),
-            }
-        };
-        let reference = match references.len() {
-            2 => Reference::Two(String::from(references[0]),
-                                String::from(references[1])),
-            1 => Reference::One(String::from(references[0])),
-            _ => Reference::None,
-        };
-        let spin = arguments.get_one::<String>("spin")
-                            .map(|x| String::from(x.as_str()));
-        Self { file,
-               file_type,
-               weight_tolerance,
-               maximum_distance,
-               output,
-               reference,
-               spin,
-               threads,
-               vacuum_tolerance }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[test]
-    fn clapapp_get() {
-        let app = ClapApp::get();
-        assert_eq!(app.get_name(), "Multi-threaded Bader Charge Analysis")
-    }
-
-    #[test]
     fn argument_file() {
-        let app = ClapApp::get();
-        let matches = app.get_matches_from(vec!["bca", "CHGCAR"]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR"];
+        let args = app.parse_args(v).unwrap();
         assert_eq!(args.file, String::from("CHGCAR"));
     }
 
     #[test]
     #[should_panic]
     fn argument_no_file() {
-        let app = ClapApp::get();
-        let _ = app.try_get_matches_from(vec!["bca"])
+        let app = App::new();
+        let _ = app.parse_args(vec!["bca"])
                    .unwrap_or_else(|e| panic!("An error occurs: {}", e));
     }
 
     #[test]
     fn argument_file_type_default_vasp() {
-        let app = ClapApp::get();
-        let matches = app.get_matches_from(vec!["bca", "CHGCAR"]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR"];
+        let args = app.parse_args(v).unwrap();
         let flag = matches!(args.file_type, FileType::Vasp);
         assert!(flag);
     }
 
     #[test]
     fn argument_file_type_default_unknown() {
-        let app = ClapApp::get();
-        let matches = app.get_matches_from(vec!["bca", "CHG"]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "CHG"];
+        let args = app.parse_args(v).unwrap();
         let flag = matches!(args.file_type, FileType::Vasp);
         assert!(flag);
     }
 
     #[test]
     fn argument_file_type_vasp() {
-        let app = ClapApp::get();
-        let matches = app.get_matches_from(vec!["bca", "CHGCAR", "-t", "vasp"]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR", "-f", "vasp"];
+        let args = app.parse_args(v).unwrap();
         let flag = matches!(args.file_type, FileType::Vasp);
         assert!(flag);
     }
 
     #[test]
     fn argument_file_type_default_cube() {
-        let app = ClapApp::get();
-        let matches = app.get_matches_from(vec!["bca", "charge.cube"]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "charge.cube"];
+        let args = app.parse_args(v).unwrap();
         let flag = matches!(args.file_type, FileType::Cube);
         assert!(flag);
     }
 
     #[test]
     fn argument_file_type_cube() {
-        let app = ClapApp::get();
-        let matches =
-            app.get_matches_from(vec!["bca", "charge.cube", "--type", "cube"]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "charge.cube", "--file_type", "cube"];
+        let args = app.parse_args(v).unwrap();
         let flag = matches!(args.file_type, FileType::Cube);
         assert!(flag);
     }
@@ -316,16 +690,16 @@ mod tests {
     #[test]
     #[should_panic]
     fn argument_file_type_not_type() {
-        let app = ClapApp::get();
-        let _ = app.try_get_matches_from(vec!["bca", "CHGCAR", "-t", "basp"])
+        let app = App::new();
+        let _ = app.parse_args(vec!["bca", "CHGCAR", "-f", "basp"])
                    .unwrap_or_else(|e| panic!("An error occurs: {}", e));
     }
 
     #[test]
     fn argument_output_atoms() {
-        let app = ClapApp::get();
-        let matches = app.get_matches_from(vec!["bca", "CHGCAR", "-o"]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR", "-o"];
+        let args = app.parse_args(v).unwrap();
         match args.output {
             WriteType::Atom(v) => assert!(v.is_empty()),
             _ => panic!(),
@@ -334,10 +708,9 @@ mod tests {
 
     #[test]
     fn argument_output_index() {
-        let app = ClapApp::get();
-        let matches =
-            app.get_matches_from(vec!["bca", "CHGCAR", "-o", "-i", "1",]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR", "-o", "-i", "1",];
+        let args = app.parse_args(v).unwrap();
         match args.output {
             WriteType::Atom(v) => assert_eq!(v, vec![0]),
             _ => panic!(),
@@ -346,10 +719,9 @@ mod tests {
 
     #[test]
     fn argument_output_mult_index() {
-        let app = ClapApp::get();
-        let matches = app.get_matches_from(vec!["bca", "CHGCAR", "-o",
-                                                "--index", "1", "-i", "3",]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR", "-o", "--index", "1", "-i", "3",];
+        let args = app.parse_args(v).unwrap();
         match args.output {
             WriteType::Atom(v) => assert_eq!(v, vec![0, 2]),
             _ => panic!(),
@@ -358,76 +730,60 @@ mod tests {
 
     #[test]
     #[should_panic]
-    fn argument_index_zero() {
-        let app = ClapApp::get();
-        let _ = app.try_get_matches_from(vec!["bca", "CHGCAR", "-o", "-i",
-                                              "0"])
-                   .unwrap_or_else(|e| panic!("An error occurs: {}", e));
-    }
-
-    #[test]
-    #[should_panic]
     fn argument_index_no_output() {
-        let app = ClapApp::get();
-        let _ = app.try_get_matches_from(vec!["bca", "CHGCAR", "-i", "1"])
+        let app = App::new();
+        let _ = app.parse_args(vec!["bca", "CHGCAR", "-i", "1"])
                    .unwrap_or_else(|e| panic!("An error occurs: {}", e));
     }
 
     #[test]
     #[should_panic]
     fn argument_index_not_parse() {
-        let app = ClapApp::get();
-        let _ = app.try_get_matches_from(vec!["bca", "CHGCAR", "-i", "1,6"])
+        let app = App::new();
+        let _ = app.parse_args(vec!["bca", "CHGCAR", "-oi", "1,6"])
                    .unwrap_or_else(|e| panic!("An error occurs: {}", e));
     }
 
     #[test]
     fn argument_spin() {
-        let app = ClapApp::get();
-        let matches = app.get_matches_from(vec!["bca",
-                                                "density.cube",
-                                                "-s",
-                                                "spin.cube",]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "density.cube", "-s", "spin.cube",];
+        let args = app.parse_args(v).unwrap();
         assert_eq!(args.spin, Some(String::from("spin.cube")))
     }
 
     #[test]
     fn argument_reference_one() {
-        let app = ClapApp::get();
-        let matches =
-            app.get_matches_from(vec!["bca", "CHGCAR", "-r", "CHGCAR_sum"]);
-        let args = Args::new(matches);
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR", "-r", "CHGCAR_sum"];
+        let args = app.parse_args(v).unwrap();
         let flag = matches!(args.reference, Reference::One(_));
         assert!(flag)
     }
 
     #[test]
     fn argument_reference_two() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "-r", "AECCAR0", "--ref", "AECCAR2"];
-        let matches = app.get_matches_from(v);
-        let args = Args::new(matches);
+        let args = app.parse_args(v).unwrap();
         let flag = matches!(args.reference, Reference::Two(_, _));
         assert!(flag)
     }
 
     #[test]
     fn argument_reference_none() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR"];
-        let matches = app.get_matches_from(v);
-        let args = Args::new(matches);
+        let args = app.parse_args(v).unwrap();
         let flag = matches!(args.reference, Reference::None);
         assert!(flag)
     }
 
     #[test]
     fn argument_aeccar() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "-a"];
-        let matches = app.get_matches_from(v);
-        let args = Args::new(matches);
+        let args = app.parse_args(v).unwrap();
         let flag = match args.reference {
             Reference::Two(x, y) => (x == *"AECCAR0") && (y == *"AECCAR2"),
             _ => false,
@@ -438,91 +794,93 @@ mod tests {
     #[test]
     #[should_panic]
     fn argument_aeccar_cube() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "charge.cube", "-a"];
-        let matches = app.get_matches_from(v);
-        let _ = Args::new(matches);
+        let _ = app.parse_args(v).unwrap();
     }
 
     #[test]
     fn argument_vacuum_tolerance_float() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "--vac", "1E-4"];
-        let matches = app.get_matches_from(v);
-        let args = Args::new(matches);
+        let args = app.parse_args(v).unwrap();
         assert_eq!(args.vacuum_tolerance, Some(1E-4))
+    }
+
+    #[test]
+    fn argument_vacuum_tolerance_default() {
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR"];
+        let args = app.parse_args(v).unwrap();
+        assert_eq!(args.vacuum_tolerance, Some(1E-6))
     }
 
     #[test]
     #[should_panic]
     fn argument_vacuum_tolerance_not_float() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "--vac", "0.00.1"];
-        let _ = app.try_get_matches_from(v)
+        let _ = app.parse_args(v)
                    .unwrap_or_else(|e| panic!("An error occurs: {}", e));
     }
 
     #[test]
     fn argument_weight_tolerance_float() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "--weight", "1E-4"];
-        let matches = app.get_matches_from(v);
-        let args = Args::new(matches);
+        let args = app.parse_args(v).unwrap();
         assert_eq!(args.weight_tolerance, 1E-4)
     }
 
     #[test]
     #[should_panic]
     fn argument_weight_tolerance_not_float() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "-w", "0.00.1"];
-        let _ = app.try_get_matches_from(v)
+        let _ = app.parse_args(v)
                    .unwrap_or_else(|e| panic!("An error occurs: {}", e));
     }
 
     #[test]
     fn argument_maximum_distance_float() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "--max_dist", "1E-4"];
-        let matches = app.get_matches_from(v);
-        let args = Args::new(matches);
+        let args = app.parse_args(v).unwrap();
         assert_eq!(args.maximum_distance, 1E-4)
     }
 
     #[test]
     #[should_panic]
     fn argument_maximum_distance_not_float() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "-m", "0.00.1"];
-        let _ = app.try_get_matches_from(v)
+        let _ = app.parse_args(v)
                    .unwrap_or_else(|e| panic!("An error occurs: {}", e));
     }
 
     #[test]
     fn argument_threads_default() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR"];
-        let matches = app.get_matches_from(v);
-        let args = Args::new(matches);
+        let args = app.parse_args(v).unwrap();
         let threads = num_cpus::get().min(12);
         assert_eq!(args.threads, threads)
     }
 
     #[test]
     fn argument_threads_int() {
-        let app = ClapApp::get();
+        let app = App::new();
         let v = vec!["bca", "CHGCAR", "--threads", "1"];
-        let matches = app.get_matches_from(v);
-        let args = Args::new(matches);
+        let args = app.parse_args(v).unwrap();
         assert_eq!(args.threads, 1)
     }
 
     #[test]
     #[should_panic]
     fn argument_threads_not_int() {
-        let app = ClapApp::get();
-        let v = vec!["bca", "CHGCAR", "-J", "0.1"];
-        let _ = app.try_get_matches_from(v)
+        let app = App::new();
+        let v = vec!["bca", "CHGCAR", "-t", "0.1"];
+        let _ = app.parse_args(v)
                    .unwrap_or_else(|e| panic!("An error occurs: {}", e));
     }
 }